Move the legacy upstart package into the service/legacy package.
